# mt-api/src/main/swagger/spec.yaml
swagger: "2.0"
info:
  title: Money Transfer Rest API
  version: "0.1"
  description: Money Transfer API specification.
host: money-transfer-api.com
basePath: /v1

schemes:
  - https

paths:

  /accounts/create:
    post:
      operationId: createAccount
      description: Create new account
      tags:
        - accounts
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/Account"
          required: true
      responses:
        201:
          description: Account created
          schema:
            required:
              - accountId
              - link
            properties:
              accountId:
                type: string
              link:
                type: string
        400:
            description: Account cant be created. See error message
            schema:
             $ref: "#/definitions/ErrorResponse"

  /accounts/{accountNumber}:
    get:
      operationId: getAccountByNumber
      description: Get Account by number
      tags:
        - accounts
      produces:
        - application/json
      parameters:
        - $ref: "#/parameters/accountNumber"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Account"
        404:
          description: Account does not exist or has been deleted

  /transfer:
    post:
      operationId: moneyTransfer
      description: Transfer funds between source and target account
      tags:
        - transaction
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/Transaction"
          required: true
      responses:
        200:
          description: Transfer successful
        400:
          description: Requested operation could not be performed. For detailed problem description refer to response body.
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Source or target account not found


parameters:
  accountNumber:
    in: path
    name: accountNumber
    description: Account Number
    type: string
    required: true


definitions:
  Account:
    type: object
    required:
      - number
    properties:
      number:
        type: string
      clientUid:
        type: string
      balance:
        $ref: "#/definitions/Amount"
      status:
        $ref: "#/definitions/AccountStatus"

  AccountStatus:
    type: string
    enum:
      - active
      - closed

  Amount:
    type: object
    required:
      - amount
      - currency
    properties:
      amount:
        type: number
      currency:
        type: string
        pattern: "[A-Z]{3}"

  Transaction:
    type: object
    required:
      - sourceAccount
      - targetAccount
      - amount
    properties:
      sourceAccount:
        type: string
      targetAccount:
        type: string
      amount:
        $ref: "#/definitions/Amount"

  ErrorResponse:
    type: object
    required:
      - code
    properties:
      code:
        description: Error code
        type: number
      message:
        description: Detailed error message
        type: string